+++ tweaks/ytmusic/src/Download/SCIYTMDownloadsController.h
@@ -1,0 +1,13 @@
+//
+//  SCIYTMDownloadsController.h
+//  Albrhi for YouTube Music
+//
+//  The saved tracks, on a screen of their own.
+//
+//  Copyright (C) Ibrahim Ismail AL-Rahn. GPLv3.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SCIYTMDownloadsController : UITableViewController
+@end